Working version; going to try to implement SQL

+++ BestLineAward/app/src/main/java/com/example/chrisc/bestlineaward/Members.java
@@ -3,24 +3,24 @@ package com.example.chrisc.bestlineaward;
 
 public class Members {
     String name;
-    double leadership;
-    double service;
-    double fellowship;
-    double total;
+    Double leadership;
+    Double service;
+    Double fellowship;
+    Double total;
     String status;
 
 
     public Members()
     {
         name = "";
-        leadership = 0;
+        leadership = 0.0;
         service = 0.0;
-        fellowship = 0;
-        total = 0;
+        fellowship = 0.0;
+        total = 0.0;
         status = "";
     }
 
-    public Members(String n, int l, double s, int f, double t, String st)
+    public Members(String n, Double l, Double s, Double f, Double t, String st)
     {
         name = n;
         leadership = l;
@@ -35,22 +35,22 @@ public class Members {
         name = n;
     }
 
-    public void setLeadership(double l)
+    public void setLeadership(Double l)
     {
         leadership = l;
     }
 
-    public void setService(double s)
+    public void setService(Double s)
     {
         service = s;
     }
 
-    public void setFellowship(double f)
+    public void setFellowship(Double f)
     {
         fellowship = f;
     }
 
-    public void setTotal(double t)
+    public void setTotal(Double t)
     {
         total = t;
     }
@@ -64,19 +64,19 @@ public class Members {
     {
         return name;
     }
-    public double getLeadership()
+    public Double getLeadership()
     {
         return leadership;
     }
-    public double getService()
+    public Double getService()
     {
         return service;
     }
-    public double getFellowship()
+    public Double getFellowship()
     {
         return fellowship;
     }
-    public double getTotal()
+    public Double getTotal()
     {
         return total;
     }
